chore(Structure): Break things into a shared crate when they are shared.
