+++ native_extensions/emoji_selector/src/style.css
@@ -33,20 +33,25 @@
 }
 
 #emoji_selector .emoji {
-  width: 25px;
-  height: 25px;
+  width: calc(var(--height-input) / 1.4);
+  height: calc(var(--height-input) / 1.4);
   display: inline-flex;
   justify-content: center;
   align-items: center;
-  font-size: 1.5rem;
+  font-size: calc(var(--height-input) / 1.4);
 }
 
 #emoji_selector .nav {
   overflow-x: scroll;
-  overflow-y: hidden;
   height: fit-content;
   width: 100%;
   gap: var(--gap);
+  padding: var(--gap-less);
+}
+
+#emoji_selector .label {
+  width: 100%;
+  display: inline-block;
   padding: var(--gap-less);
 }
