+++ shared/Cargo.toml
@@ -10,3 +10,36 @@ unic-langid = "0.9.1"
 isolang = "2.2.0"
 fluent-templates = "0.8.0"
 warp = { git = "https://github.com/Satellite-im/Warp", rev = "a7e075238ee926919894cb27d743284ed3724dc2" }
+notify = "5.1.0"
+serde = "1.0.150"
+serde_json = "1.0.89"
+once_cell = "1.13"
+clap = { version = "4.0.32", features = ["derive"] }
+rodio = "0.16"
+uuid = { version = "1.0", features = ["serde", "v4"] }
+notify-rust = { version = "4.6.0", default-features = false, features = ["d"] }
+dirs = "4.0.0"
+base64 = "0.20.0"
+timeago = "0.4.0"
+lipsum = "0.8.2"
+chrono = "0.4.19"
+rand = "0.8.4"
+image = "0.24.5"
+names = "0.14.0"
+titlecase = "2.2.1"
+substring = "1.4.5"
+extensions = { path = "../extensions" }
+icons = { path="../icons" }
+
+dioxus = { git = "https://github.com/DioxusLabs/dioxus", rev="d521da1991719760e271457dfe4f9ddf281afbb3" }
+dioxus-desktop = { git = "https://github.com/DioxusLabs/dioxus", rev = "d521da1991719760e271457dfe4f9ddf281afbb3", features = [
+    "transparent",
+] }
+derive_more = "0.99"
+either = "1.6.1"
+wry = { version = "0.23.4" }
+
+[target.'cfg(target_os = "macos")'.dependencies]
+cocoa = "0.24.1"
+objc = "0.2.7"
+core-foundation = "0.9.3"
