+++ shared/src/state/configuration.rs
