+++ shared/src/state/friends.rs
