+++ shared/src/state/settings.rs
