+++ ui/Cargo.toml
@@ -21,7 +21,6 @@ tokio-util = { version = "0.7", features = ["full"] }
 arboard = "3.1.0"
 humansize = "2.1.3"
 uuid = { version = "1.0", features = ["serde", "v4"] }
-lipsum = "0.8.2"
 libloading = "0.7.4"
 warp = { git = "https://github.com/Satellite-im/Warp", rev = "a7e075238ee926919894cb27d743284ed3724dc2" }
 warp-mp-ipfs = { git = "https://github.com/Satellite-im/Warp", rev = "a7e075238ee926919894cb27d743284ed3724dc2" }
@@ -29,20 +28,14 @@ warp-rg-ipfs = { git = "https://github.com/Satellite-im/Warp", rev = "a7e075238e
 warp-fs-ipfs = { git = "https://github.com/Satellite-im/Warp", rev = "a7e075238ee926919894cb27d743284ed3724dc2" }
 rfd = "0.10.0"
 mime = "0.3.16"
-names = "0.14.0"
-titlecase = "2.2.1"
 serde = "1.0.150"
 serde_json = "1.0.89"
-rand = "0.8.4"
-image = "0.24.5"
 chrono = "0.4.19"
-substring = "1.4.5"
 either = "1.6.1"
 base64 = "0.20.0"
 timeago = "0.4.0"
 dirs = "4.0.0"
 opener = "0.5.0"
-rodio = "0.16"
 open = "3.2.0"
 wry = { version = "0.23.4" }
 derive_more = "0.99"
@@ -52,7 +45,6 @@ notify = "5.1.0"
 tempfile = "3.0.7"
 fdlimit = "0.2"
 
-notify-rust = { version = "4.6.0", default-features = false, features = ["d"] }
 once_cell = "1.13"
 
 futures = "0.3"
